api/buzzer: fix tone duration being a thousand times too long
BUZZER_tone treated its duration parameter as seconds and multiplied by
1000, but every caller passes milliseconds. BUZZER_Play_Pirates derives its
note lengths as 1000 / duration and pairs them with HAL_Delay, confirming
the intended unit.

On the hardware this made a single button handler busy-wait for three
consecutive 150 second tones. The main loop stalled for exactly seven and
a half minutes after boot, with SysTick still running, so the display froze
on its first frame while uwTick kept advancing.

Rename the parameter and drop the multiplier. Widen it to uint32_t; the
Pirates tune already generates 250, which barely fits a uint8_t.

+++ app/hot_plate/hot_place.cpp
@@ -98,7 +98,7 @@ void loop() {
         u8g2_DrawStr(&u8g2,20, 10,"    COMPLETE   ");
       } while ( u8g2_NextPage(&u8g2) );
 
-      BUZZER_tone((buzzer_t *)&buzzer1, (uint32_t)1800, (uint8_t)100);
+      BUZZER_tone((buzzer_t *)&buzzer1, (uint32_t)1800, (uint32_t)100);
       seconds = 0;              //Reset timer
       running_mode = hotPlateState_COOLDOWN;
       delay(3000);
